+++ tests/conftest.py
@@ -19,8 +19,23 @@ def check(output: str) -> bool:
         with open(compare_file, "w") as outfile:
             outfile.write(yaml_output)
     else:
-        with open(compare_file, "r") as infile:
-            assert yaml_output == infile.read()
+        try:
+            with open(compare_file, "r") as infile:
+                assert yaml_output == infile.read()
+        except FileNotFoundError as exc:
+            print(
+                "Comparison file not found.",
+                "Create it by executing:\n\n",
+                f"\tUPDATE_TEST_OUTPUT=true pytest {caller_file_path}",
+            )
+            raise exc
+        except AssertionError as exc:
+            print(
+                "If intentionally, you can update the comparions file:\n\n",
+                "\trm -rf test/unit/comparison_files/*; UPDATE_TEST_OUTPUT=true pytest\n\n",
+                "Always review the results carefully with 'git diff'!",
+            )
+            raise exc
 
 
 @pytest.fixture